# web/src/lib/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gastify API",
    "description": "Chilean smart expense tracker \u2014 FastAPI backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/scans": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Submit Scan",
        "operationId": "submit_scan_api_v1_scans_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_scan_api_v1_scans_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanSubmission"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{scan_id}/process": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Trigger Process Scan",
        "description": "Manually trigger extraction for a scan (retry or reprocess).",
        "operationId": "trigger_process_scan_api_v1_scans__scan_id__process_post",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/scan-test-cases/{case_id}/runs": {
      "post": {
        "tags": [
          "scan-test-cases"
        ],
        "summary": "Run Test Case",
        "operationId": "run_test_case_api_v1_scan_test_cases__case_id__runs_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanTestRunSubmission"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_submit_scan_api_v1_scans_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_submit_scan_api_v1_scans_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScanResult": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "processed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "submitted_at"
        ],
        "title": "ScanResult"
      },
      "ScanSubmission": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ownership_scope_id": {
            "type": "string",
            "format": "uuid",
            "title": "Ownership Scope Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "file_size_bytes": {
            "type": "integer",
            "title": "File Size Bytes"
          },
          "image_path": {
            "type": "string",
            "title": "Image Path"
          },
          "thumbnail_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Path"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ownership_scope_id",
          "status",
          "original_filename",
          "content_type",
          "file_size_bytes",
          "image_path",
          "submitted_at"
        ],
        "title": "ScanSubmission"
      },
...
      "ScanTestRunSubmission": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ownership_scope_id": {
            "type": "string",
            "format": "uuid",
            "title": "Ownership Scope Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "file_size_bytes": {
            "type": "integer",
            "title": "File Size Bytes"
          },
          "image_path": {
            "type": "string",
            "title": "Image Path"
          },
          "thumbnail_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Path"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "test_case_id": {
            "type": "string",
            "title": "Test Case Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "convenience_only": {
            "type": "boolean",
            "title": "Convenience Only"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ownership_scope_id",
          "status",
          "original_filename",
          "content_type",
          "file_size_bytes",
          "image_path",
          "submitted_at",
          "test_case_id",
          "provider",
          "convenience_only"
        ],
        "title": "ScanTestRunSubmission"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}